feat: Refactor To Domain / Infra Layer
