+++ src/main/java/com/commerce/backendserver/auth/application/CustomOAuth2LoginService.java
@@ -4,8 +4,8 @@ import com.commerce.backendserver.auth.infra.oauth.OAuthProfile;
 import com.commerce.backendserver.auth.infra.oauth.OAuthProfileFactory;
 import com.commerce.backendserver.auth.infra.oauth.UserPrincipal;
 import com.commerce.backendserver.member.domain.Member;
-import com.commerce.backendserver.member.domain.MemberQueryRepository;
-import com.commerce.backendserver.member.domain.MemberRepository;
+import com.commerce.backendserver.member.infra.persistence.MemberQueryRepository;
+import com.commerce.backendserver.member.infra.persistence.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
