+++ src/main/java/com/commerce/backendserver/global/config/SecurityConfig.java
@@ -3,12 +3,11 @@ package com.commerce.backendserver.global.config;
 import com.commerce.backendserver.auth.application.filter.JwtAuthenticationFilter;
 import com.commerce.backendserver.auth.application.filter.JwtExceptionHandlerFilter;
 import com.commerce.backendserver.auth.infra.jwt.JwtProvider;
-import com.commerce.backendserver.member.domain.MemberQueryRepository;
+import com.commerce.backendserver.member.infra.persistence.MemberQueryRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
@@ -27,7 +26,6 @@ import org.springframework.web.cors.CorsConfiguration;
 import java.util.Arrays;
 import java.util.Collections;
 
-import static java.lang.invoke.VarHandle.AccessMode.*;
 import static org.springframework.http.HttpMethod.GET;
 
 @Configuration
