+++ src/main/java/com/commerce/backendserver/member/infra/persistence/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.commerce.backendserver.member.domain;
+package com.commerce.backendserver.member.infra.persistence;
 
+import com.commerce.backendserver.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
