--- src/main/java/com/commerce/backendserver/product/domain/infra/query/ProductQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.commerce.backendserver.product.domain.infra.query;
-
-public interface ProductQueryDslRepository {
-}
