+++ src/main/java/com/commerce/backendserver/product/infra/persistence/ProductJpaRepository.java
@@ -1,4 +1,4 @@
-package com.commerce.backendserver.product.domain.persistence;
+package com.commerce.backendserver.product.infra.persistence;
 
 import com.commerce.backendserver.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
