+++ src/main/java/com/commerce/backendserver/product/infra/persistence/ProductQueryDslRepository.java
@@ -1,0 +1,4 @@
+package com.commerce.backendserver.product.infra.persistence;
+
+public interface ProductQueryDslRepository {
+}
