+++ src/main/java/com/commerce/backendserver/product/infra/persistence/promotion/PromotionJpaRepository.java
@@ -1,4 +1,4 @@
-package com.commerce.backendserver.product.domain.persistence.promotion;
+package com.commerce.backendserver.product.infra.persistence.promotion;
 
 import com.commerce.backendserver.product.domain.promotion.Promotion;
 import org.springframework.data.jpa.repository.JpaRepository;
