+++ src/main/java/com/commerce/backendserver/review/infra/ReviewRepository.java
@@ -1,5 +1,6 @@
-package com.commerce.backendserver.review.domain;
+package com.commerce.backendserver.review.infra;
 
+import com.commerce.backendserver.review.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
