+++ src/test/java/com/commerce/backendserver/review/application/ReviewServiceTest.java
@@ -4,10 +4,10 @@ import com.commerce.backendserver.common.base.MockTestBase;
 import com.commerce.backendserver.global.exception.CommerceException;
 import com.commerce.backendserver.image.application.ImageService;
 import com.commerce.backendserver.product.domain.Product;
-import com.commerce.backendserver.product.domain.persistence.ProductQueryRepository;
+import com.commerce.backendserver.product.infra.persistence.ProductQueryRepository;
 import com.commerce.backendserver.review.application.dto.request.CreateReviewRequest;
 import com.commerce.backendserver.review.domain.Review;
-import com.commerce.backendserver.review.domain.ReviewRepository;
+import com.commerce.backendserver.review.infra.ReviewRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
